extension: build against tree-sitter
The whole idea of putting tree-sitter as a submodule is a major mistake
for many reasons:

1. It's a pain in the ass to make ruby compile the bindings against the
   submodule's lib, let alone linking statically.
   1. Information on how to do it properly on the interwebs are scarce,
      and I am starting to guess that a lot of people don't do it
      because it's such a time-sink
   1. Statically linking to the generated output? GG
1. Even if you make it compile, I encountered a lot of issues.  `mkmf`
   will happily compile and link with 0 complaints, except:
   1. When you try to run a REPL and `require 'tree_sitter'`: symbols
      are yet to be found.
   1. When you try to build a `.gem` and install it. It will not find
      the sources you're looking for because, well, they're in a
      submodule, and its content are not packaged by bundler. Making
      this work is another time-sink that didn't even consider.

So for now, we're going to rely on system-installed tree-sitter.
We're doing a lookup in the usual places (/usr /opt).

If someone wants to work with specific versions of the lib, they'll have
to either pass some `--with-tree-sitte-*` options to `bundle config`
before building the gem, or they'll have to do some docker shenanigans
and work there.