node: add is_null, and `null?` in ruby
In order to distinguish it from Ruby's own `nil?` I have decided to call
it `null?`.

As a matter of fact, `ts_node_is_null` doesn't check agains NULL.
Tree-Sitter considers that ts_node_is_null := true iff node.id = 0.
That might be subject to changes, since it's in the .c files, but this
is where it stands as of now.