query_cursor: add next_capture
I suspect that this implementation is highly inefficient because this
method is intended to be used in a loop.

So if we're creating an array of two elements and returning them every
single time, this might affect performance (intuitively speaking of
course).

This puts in question how thin the binding need to be.