range: make TypedData
This is needed because ranges need to be used in ruby-space.
Using the APIs that return ranges must return a valid ruby object,
and we need to avoid conversion everytime we access any of the object
members.

Any slowdown due to conversion from range_t to TSRange while using the
direct wrappers is unavoidable.

More performant operations should use TSRange directly, and these will
be implemented in C, and not ruby.