docs: notes: we don't have copying issues anymore
Since we moved to the WRAP macros, we keep the data as inside ruby-space
and we convert them to C-space when needed.

So the issue is a bit different now, so I'd rather remove the text and
add meaningful explanation when the library is stable and when we do
some benchmarks.