node: fix method_missing impl
This is not perfect. There's a bug in tree-sitter where present
field_name would not be listed when iterating through a node's child.
It's in the node's structural_index but not child_index, i.e.
it's in the children of the node.

Which means that the results of any access to field_name is not
guaranteed.