extconf: enable download from source + static linking
Which is mostly for security-oriented people who don't want
to use git and curl to download external dependencies, then
open unwanted secrity risks.

Yes this build process will use git or curl, but the end-user
will be using a pre-package, platform-specific gem, therefore
they'll only have a single point of entry to the outside world:
the sources of their gemfiles.

I personally dislike this kind of approaches, but to each his own.