feat!: Integrate TreeStand
The [TreeStand](https://github.com/Shopify/tree_stand/issues/30) project
offers a higher-level abstraction on top of the tree-sitter bindings
provided by this project. It was decided in
https://github.com/Shopify/tree_stand/issues/30 that the API be
integrated directly into ruby-tree-sitter. This commit integrates the
TreeStand API & it's testsuite into the rest of ruby-tree-sitter.

The TreeStand API is kept in a separate structure so it can be loaded
separately from the original ruby bindings.

BREAKING CHANGE: drop support for ruby 2.7