test: do not include the debug gem by default
When the test suite fails it opens an rdbg session, and most of the time,
when debugging, I'd like to just get back to my code and work with it.

It would have been OK if a ^C or a ^D would do it. But the problem is
worst: my terminal keeps getting out of sync.

Uncomment that line if you want that debugging functionality. It's now
opt-in.