ci & rake: disable native gem compilation
I'm either missing on something in the docs of rake-compiler and
rake-compiler-dock, or the doc is missing some crucial information.

Trying to read the existing repos is not helping that much neither
(especially Nokogiri).

If I find the opportunity to spend all the time necessary to fix this, I
would gladly spend it. If someone else would like to contribute, it will
be even better, because I don't think it's possible in the near future.