extconf: dir_config on the correct directory
Fixes the following bug:

When tree-sitter is installed in the system, --disable-sys-libs didn't
actually statically link to the downloaded tree-sitter, and we used
to resolve to the system installation, thus preferring the dylib/so
over the a.

I am still having an issue forcing static compilation without needing
to delete the dylib built from the downloaded tree-sitter.