mixin: lang: be more leniant in language loading.
This is mostly relevant to times where you need to load `COBOL` for
example. Now we can just say `TreeSitter.load('COBOL')`, and that will
be enough to load correctly the symbol 'COBOL' from the compiler parser,
and at the same capable of looking up parsers
`cobol.so/COBOL.so/tree-sitter-cobol.so/tree-sitter-COBOL.so/…`.